+++ examples/shop-with-blog/client/src/pages/shop/Category/Category.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { NetworkStatus } from 'apollo-client';
-import { Helmet } from 'react-helmet-async';
 import { Toggle } from 'react-powerplug';
 import { CategoryWithProductListQuery } from '@deity/falcon-shop-data';
 import { H1, Box, FlexLayout, Divider, Button } from '@deity/falcon-ui';
-import { SearchConsumer, SortOrderPickerProvider, getFiltersData } from '@deity/falcon-front-kit';
+import { SearchConsumer, SortOrderPickerProvider, SEO, getFiltersData } from '@deity/falcon-front-kit';
 import {
   CategoryLayout,
   CategoryArea,
@@ -38,9 +37,7 @@ const CategoryPage = ({ match: { params } }) => (
 
           return (
             <CategoryLayout variant={!filtersData.length && 'noFilters'}>
-              <Helmet>
-                <title>{name}</title>
-              </Helmet>
+              <SEO meta={category.seo} title={name} />
               <Box gridArea={CategoryArea.heading}>
                 <H1>{name}</H1>
                 <FlexLayout justifyContent="space-between" alignItems="center">