+++ CHANGELOG.md
@@ -56,6 +56,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 - updated Formik to version 2.0 ([#705](https://github.com/deity-io/falcon/pull/705))
 - added `InBrowserOnly` component ([#728](https://github.com/deity-io/falcon/pull/728))
 - added `SwitchDynamicURL` component in order to define dynamic URL routes in `react`/`react-router` `v4` way ([#738](https://github.com/deity-io/falcon/pull/738))
+- added `SEO` component to render `title`, `description` and `keywords` as proper meta-tags ([#743](https://github.com/deity-io/falcon/pull/743))
 
 ### Falcon UI Kit v1.0.0
 
@@ -89,6 +90,8 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 ### Falcon Shop Extension v1.0.0
 
 - introduced Country and Region types to allow fetching more details with address queries ([#609](https://github.com/deity-io/falcon/pull/609))
+- added `@cache` for `Query.productList` query ([#743](https://github.com/deity-io/falcon/pull/743))
+- added `Product.categories` and `Category.{urlPath,image,attributes,seo}` fields ([#743](https://github.com/deity-io/falcon/pull/743))
 
 ### Falcon Server vNext